+++ Gears/Misc.h
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace Misc {
+    void UpdateEngineOnOff();
+}